Moved elastic search tests to Fixture
